created the 'Move' component
