+++ react-app/src/app/screens/Game/components/Move/styles.scss
@@ -1,0 +1,12 @@
+@import '../../../../../scss/colors';
+
+.move {
+  background: $white;
+  border: 1px solid $dusty-gray;
+  border-radius: 10px;
+
+  &:hover{
+    background: $mine-shaft;
+    color: $white;
+  }
+}
