+++ react-app/src/app/screens/Game/components/Moves/styles.scss
@@ -1,12 +1,1 @@
 @import '../../../../../scss/colors';
-
-.move {
-  background: $white;
-  border: 1px solid $dusty-gray;
-  border-radius: 10px;
-
-  &:hover{
-    background: $mine-shaft;
-    color: $white;
-  }
-}
